data_engineering/pyproject.toml
[project]
name = "weather-inmet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Alexandre-Esposte", email = "alexandreesposte@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "unidecode>=1.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
